requirements.txt
# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0

# Configuration and data handling
pyyaml>=6.0
pydantic>=2.5.0
click>=8.1.0

# Notifications
python-telegram-bot>=20.7

# Scheduling and async
schedule>=1.2.0
asyncio-mqtt>=0.13.0

# Logging and utilities
colorlog>=6.7.0
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
